Add common interface for tool path planners
