--- noether/config/halfedge_boundary_finder.yaml
@@ -1,6 +1,0 @@
-min_num_points: 50
-point_spacing_method: 2 # NONE= 0, MIN_DISTANCE = 1, EQUAL_SPACING = 2, PARAMETRIC_SPLINE = 3 
-point_dist: 0.02
-normal_averaging: True
-normal_search_radius: 0.1
-normal_influence_weight: 1.0
