+++ noether/include/noether/surface_raster_planner_application.h
@@ -24,7 +24,7 @@
 #include <vtk_viewer/vtk_utils.h>
 #include <mesh_segmenter/mesh_segmenter.h>
 #include <path_sequence_planner/path_sequence_planner.h>
-#include <tool_path_planner/raster_tool_path_planner.h>
+#include <tool_path_planner/surface_walk_raster_generator.h>
 
 namespace noether
 {
