--- noether_msgs/action/GenerateEdgePaths.action
@@ -1,12 +1,0 @@
-# Goal
-shape_msgs/Mesh[] surface_meshes  # The surfaces on which to generate paths.
-bool proceed_on_failure           # When True, edge generation will proceed to the next surface when an error is encountered. 
----
-# Result
-bool success
-string err_msg
-ToolRasterPath[] edge_paths   # The resulting raster paths
-bool[] validities             # True when the edge path in 'tool_raster_paths' with the same index is valid.
----
-# Feedback
-int32 current_mesh_index        # The index of the mesh being processed
