+++ noether_msgs/msg/ToolPath.msg
@@ -1,0 +1,1 @@
+geometry_msgs/PoseArray[] segments  # The tool path segments
