+++ noether_msgs/msg/ToolPathConfig.msg
@@ -1,17 +1,12 @@
-float64 pt_spacing 					# Required spacing between path points
-float64 line_spacing 				# Offset between two rasters
-float64 tool_offset 				# How far off the surface the tool needs to be
-float64 intersecting_plane_height 	# Used by the raster_tool_path_planner when offsetting to an adjacent path, a new plane has to be formed, but not too big
-float64 min_hole_size 				# A path may pass over holes smaller than this, but must be broken when larger holes are encounterd. 
-float64 min_segment_size 			# The minimum segment size to allow when finding intersections; small segments will be discarded
-float64 raster_angle          # Specifies the direction of the raster path in radians
+uint8 SURFACE_WALK_RASTER_GENERATOR=0
+uint8 PLANE_SLICER_RASTER_GENERATOR=1
+uint8 HEAT_RASTER_GENERATOR=2
+uint8 EIGEN_VALUE_EDGE_GENERATOR=3 # This does not have a config
+uint8 HALFEDGE_EDGE_GENERATOR=4
 
-# When set to TRUE: initial cuts are determined using the axes of the
-# coordinate frame in which the mesh is placed.  This may cause the
-# initial cut to miss the part.
-# When set to FALSE: initial cuts are determined using the centroid and
-# principal axes of the part.  This is the old default behavior.  (Note
-# that ROS sets message fields to 0/False by default.)
-bool raster_wrt_global_axes
-
-bool generate_extra_rasters   # Whether an extra raster stroke should be added to the beginning and end of the raster pattern
+uint8 type
+noether_msgs/SurfaceWalkRasterGeneratorConfig surface_walk_generator
+noether_msgs/PlaneSlicerRasterGeneratorConfig plane_slicer_generator
+#noether_msgs/HeatRasterPathGeneratorConfig heat_generator
+noether_msgs/EigenValueEdgeGeneratorConfig eigen_value_generator
+noether_msgs/HalfedgeEdgeGeneratorConfig halfedge_generator
