+++ noether_msgs/msg/ToolPaths.msg
@@ -1,0 +1,1 @@
+noether_msgs/ToolPath[] paths  # The tool paths, include surface or edge
