--- noether_msgs/msg/ToolRasterPath.msg
@@ -1,1 +1,0 @@
-geometry_msgs/PoseArray[] paths			# The raster paths on a surface
