+++ tool_path_planner/include/tool_path_planner/plane_slicer_raster_generator.h
@@ -28,25 +28,82 @@
 #include <boost/optional.hpp>
 #include <pcl/PolygonMesh.h>
 #include <shape_msgs/Mesh.h>
-#include <geometry_msgs/PoseArray.h>
 #include <vtkCellLocator.h>
 #include <vtkKdTreePointLocator.h>
-#include <noether_msgs/ToolRasterPath.h>
+#include <jsoncpp/json/json.h>
+#include <ros/console.h>
+
+#include <tool_path_planner/path_generator.h>
 
 namespace tool_path_planner
 {
-class PlaneSlicerRasterGenerator
+class PlaneSlicerRasterGenerator : public PathGenerator
 {
+  static constexpr double DEFAULT_RASTER_SPACING = 0.04;
+  static constexpr double DEFAULT_POINT_SPACING = 0.01;
+  static constexpr double DEFAULT_RASTER_ROT_OFFSET = 0.0;
+  static constexpr double DEFAULT_MIN_SEGMENT_SIZE = 0.01;
+  static constexpr double DEFAULT_SEARCH_RADIUS = 0.01;
+  static constexpr double DEFAULT_MIN_HOLE_SIZE = 1e-2;
+
 public:
 
   struct Config
   {
-    double raster_spacing = 0.04;
-    double point_spacing = 0.01;
-    double raster_rot_offset = 0.0;
-    double min_segment_size = 0.01;
-    double search_radius = 0.01;
-    double min_hole_size = 1e-2;
+    double raster_spacing {DEFAULT_RASTER_SPACING};
+    double point_spacing {DEFAULT_POINT_SPACING};
+    double raster_rot_offset {DEFAULT_RASTER_ROT_OFFSET};
+    double min_segment_size {DEFAULT_MIN_SEGMENT_SIZE};
+    double search_radius {DEFAULT_SEARCH_RADIUS};
+    double min_hole_size {DEFAULT_MIN_HOLE_SIZE};
+
+    Json::Value toJson() const
+    {
+      Json::Value jv(Json::ValueType::objectValue);
+      jv["raster_spacing"] = raster_spacing;
+      jv["point_spacing"] = point_spacing;
+      jv["raster_rot_offset"] = raster_rot_offset;
+      jv["min_segment_size"] = min_segment_size;
+      jv["search_radius"] = search_radius;
+      jv["min_hole_size"] = min_hole_size;
+
+      return jv;
+    }
+
+    bool fromJson(const Json::Value& jv)
+    {
+      if(jv.isNull())
+      {
+        ROS_ERROR("Json value is null");
+        return false;
+      }
+      if(jv.type() != Json::ValueType::objectValue)
+      {
+        ROS_ERROR( "Json type %i is invalid, only '%i' is allowed",static_cast<int>(jv.type()), static_cast<int>(Json::ValueType::objectValue));
+        return false;
+      }
+      auto validate = [](const Json::Value& jv,const std::string& name_, const Json::ValueType& type_) -> bool
+      {
+        return jv.isMember(name_) && jv[name_].type() == type_;
+      };
+      raster_spacing = validate(jv,"raster_spacing",Json::ValueType::realValue) ? jv["raster_spacing"].asDouble() : DEFAULT_RASTER_SPACING;
+      point_spacing = validate(jv,"point_spacing",Json::ValueType::realValue) ? jv["point_spacing"].asDouble() : DEFAULT_POINT_SPACING;
+      raster_rot_offset = validate(jv,"raster_rot_offset", Json::ValueType::realValue) ? jv["raster_rot_offset"].asDouble() : DEFAULT_RASTER_ROT_OFFSET;
+      min_segment_size = validate(jv,"min_segment_size", Json::ValueType::realValue) ? jv["min_segment_size"].asDouble() : DEFAULT_MIN_SEGMENT_SIZE;
+      search_radius = validate(jv,"search_radius", Json::ValueType::realValue) ? jv["search_radius"].asDouble() : DEFAULT_SEARCH_RADIUS;
+      min_hole_size = validate(jv,"min_hole_size",Json::ValueType::realValue) ? jv["min_hole_size"].asDouble() : DEFAULT_MIN_HOLE_SIZE;
+      return true;
+    }
+
+    bool fromJson(const std::string& jv_string)
+    {
+      Json::Value jv;
+      Json::Reader r;
+      if (!r.parse(jv_string, jv))
+        return false;
+
+      return fromJson(jv);
+    }
 
     std::string str()
     {
@@ -60,51 +117,27 @@ public:
     }
   };
 
-  PlaneSlicerRasterGenerator();
-  virtual ~PlaneSlicerRasterGenerator();
-
-
-  /**
-   * @brief sets the input mesh from which raster paths are to be generated
-   * @param mesh The mesh input
-   */
-  void setInput(pcl::PolygonMesh::ConstPtr mesh);
+  PlaneSlicerRasterGenerator() = default;
+  virtual ~PlaneSlicerRasterGenerator() = default;
 
   /**
-   * @brief sets the input mesh from which raster paths are to be generated
-   * @param mesh The mesh input
-   */
-  void setInput(const shape_msgs::Mesh& mesh);
-
-  /**
-   * @brief Generate the raster paths that follow the contour of the mesh
+   * @brief Set the generator configuration
    * @param config The configuration
-   * @return  An array of raster paths or boost::none when it fails.
+   * @return True if valid configuration, otherwise false.
    */
-  boost::optional< std::vector<noether_msgs::ToolRasterPath> > generate(const PlaneSlicerRasterGenerator::Config& config);
+  void setConfiguration(const Config& config);
 
-  /**
-   * @brief Generate the raster paths that follow the contour of the mesh
-   * @param mesh  The input mesh from which raster paths will be generated
-   * @param config The configuration
-   * @return  An array of raster paths or boost::none when it fails.
-   */
-  boost::optional< std::vector<noether_msgs::ToolRasterPath> > generate(const shape_msgs::Mesh& mesh,
-                                                                    const PlaneSlicerRasterGenerator::Config& config);
+  void setInput(pcl::PolygonMesh::ConstPtr mesh) override;
 
-  /**
-   * @brief Generate the raster paths that follow the contour of the mesh
-   * @param mesh  The input mesh from which raster paths will be generated
-   * @param config The configuration
-   * @return  An array of raster paths or boost::none when it fails.
-   */
-  boost::optional< std::vector<noether_msgs::ToolRasterPath> > generate(pcl::PolygonMesh::ConstPtr mesh,
-                                                                    const PlaneSlicerRasterGenerator::Config& config);
-  /**
-   * @brief the class name
-   * @return a string
-   */
-  std::string getName();
+  void setInput(vtkSmartPointer<vtkPolyData> mesh) override;
+
+  void setInput(const shape_msgs::Mesh& mesh) override;
+
+  vtkSmartPointer<vtkPolyData> getInput() override;
+
+  boost::optional<ToolPaths> generate() override;
+
+  std::string getName() const override;
 
 private:
 
@@ -113,6 +146,7 @@ private:
   vtkSmartPointer<vtkPolyData> mesh_data_;
   vtkSmartPointer<vtkKdTreePointLocator> kd_tree_;
   vtkSmartPointer<vtkCellLocator> cell_locator_;
+  Config config_;
 };
 
 } /* namespace tool_path_planner */
